iamodel/requirements_model.txt
tensorflow==2.15.*
Pillow
numpy
Flask